Add to_regtypemod() SQL function
The `to_regtypemod()` function uses the underlying `parseTypeString()` C
function to parse a string representing a data type into a type ID
(which id discards) and a typmod, which it returns. This value is
suitable for passing as the second argument to `format_type()`, which
allow one to derive the formal SQL typmod from a string:

    SELECT format_type(
      to_regtype('varchar(32)'),
      to_regtypemod('varchar(32)')
    );
          format_type
    -----------------------
     character varying(32)

This function also resolves types whose typmod is determined by the SQL
parser or some step after that, such as interval types where the stored
field option is encoded in the typmod:

    SELECT format_type(
      to_regtype('interval second(0)'),
      to_regtypemod('interval second(0)')
    );
        format_type
    --------------------
    interval second(0)

Useful for unit tests for against column data types, for example. Based
on code originally written by Erik Wienhold for use in pgTAP.

Like `to_regtype()`, `to_regtypemod()` raises an error when the
`parseTypeString()` function drops into the SQL grammar parser raises
an error. So invalid type names like `inteval` or `interval nonesuch`
raise an error, rather than return `NULL`. Note this behavior in the
docs, and also update the documentation for `to_regtype()` to note
this behavior and cite examples.

While at it, fix an "use of undefined value in sprintf" error in
`genbki.pl` that arose when I had specified `prorettype => 'integer'`
instead of `prorettype => 'int4'` in `pg_proc.dat`. I only realized the
invalid type name when I fixed the Perl bug.