Teach jsonpath string() to unwrap in lax mode
All other jsonpath methods, aside from type() and .size(), will unwrap
an array in lax mode, but string(), added in 66ea94e, overlooked this
behavior. A discussion on pgsql-hackers[1] cites the SQL standard:

> General Rule 11 g ii 6) A) says just "if MODE is lax and <JSON method>
> is not type or size, then let BASE be Unwrap(BASE)." No special
> exemption there for string(), nor further below at C) XV) for the
> operation of string().

So teach string() to also unwrap in lax mode, update the test for this
behavior, and add a line to the docs about the behavior of methods in
lax mode.

  [1]: https://www.postgresql.org/message-id/flat/A64AE04F-4410-42B7-A141-7A7349260F4D%40justatheory.com

Signed-off-by: David E. Wheeler <david@justatheory.com>